Refactor window and pyserver logic out of main.ts
