+++ src/main/window.ts
@@ -1,0 +1,63 @@
+/* eslint global-require: off, no-console: off, promise/always-return: off */
+
+import { ChildProcess } from 'child_process';
+import { app, BrowserWindow, shell } from 'electron';
+import path from 'path';
+import { startServer, pingServer } from './pyServer';
+import startExpressServer from './expressServer';
+import { isDevelopment, resolveHtmlPath } from './util';
+import installExtensions from './extensions';
+
+const createWindow = async () => {
+  if (isDevelopment) {
+    await installExtensions();
+  }
+
+  const RESOURCES_PATH = app.isPackaged
+    ? path.join(process.resourcesPath, 'assets')
+    : path.join(__dirname, '../../assets');
+
+  const getAssetPath = (...paths: string[]): string => {
+    return path.join(RESOURCES_PATH, ...paths);
+  };
+
+  const mainWindow = new BrowserWindow({
+    show: false,
+    width: 1024,
+    height: 728,
+    icon: getAssetPath('icon.png'),
+    webPreferences: {
+      preload: path.join(__dirname, 'preload.js'),
+      nodeIntegration: true,
+    },
+  });
+
+  mainWindow.loadURL(resolveHtmlPath('index.html'));
+
+  mainWindow.on('ready-to-show', async () => {
+    if (!mainWindow) {
+      throw new Error('"mainWindow" is not defined');
+    }
+    if (process.env.START_MINIMIZED) {
+      mainWindow.minimize();
+    } else {
+      mainWindow.show();
+    }
+
+    const pyServer: ChildProcess = startServer();
+
+    pingServer(pyServer);
+
+    startExpressServer();
+  });
+
+  // Open urls in the user's browser
+  mainWindow.webContents.setWindowOpenHandler((edata) => {
+    shell.openExternal(edata.url);
+    return { action: 'deny' };
+  });
+
+  return mainWindow;
+};
+
+export default createWindow;
